+++ iaw01/Scenes/VerifyPhoneNumberScene/VerifyPhoneNumberView.swift
@@ -63,8 +63,8 @@ class VerifyPhoneNumberView: UIView {
         return button
     }()
     
-    private lazy var codeInputView: CodeInputView = {
-        let view = CodeInputView()
+    private lazy var codeInputView: PincodeInputView = {
+        let view = PincodeInputView()
         view.translatesAutoresizingMaskIntoConstraints = false
         return view
     }()
